refactor(components): move `RelativeTime` and `GenreChip` to `/items`
