+++ app/components/items/index.ts
@@ -6,3 +6,5 @@ export * from './list/element.skeleton'
 export * from './image.skeleton'
 export * from './list.skeleton'
 export * from './top-card.skeleton'
+export * from './relative-time'
+export * from './genre-chip'
