+++ app/profile/[user-id]/top/genres/page.tsx
@@ -3,7 +3,8 @@ import { redirect } from 'next/navigation'
 import { validateId } from '@app/utils/validate-id'
 import { TIME_RANGE, USER_ID } from '@app/constants'
 import { validateTimeRange } from '@app/profile/utils/time-range'
-import { GenreChip, ToggleTimeRange } from '@app/components/common'
+import { ToggleTimeRange } from '@app/components/common/'
+import { GenreChip } from '@app/components/items'
 import { getTopGenres } from '@app/api/fetchers'
 import { DefaultSection } from '@app/sections'
 import { ProfilePageProps } from '@app/profile/types'
